Refactor to unify code to add CAPI and TF templates
