+++ pkg/adapters/http_test.go
@@ -14,9 +14,8 @@ import (
 	"github.com/weaveworks/weave-gitops/pkg/adapters"
 	"github.com/weaveworks/weave-gitops/pkg/capi"
 	"github.com/weaveworks/weave-gitops/pkg/clusters"
+	"github.com/weaveworks/weave-gitops/pkg/testutils"
 )
-
-const BaseURI = "https://weave.works/api"
 
 func TestRetrieveTemplates(t *testing.T) {
 	tests := []struct {
@@ -68,9 +67,9 @@ func TestRetrieveTemplates(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/templates", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/templates", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			ts, err := r.RetrieveTemplates()
 			tt.assertFunc(t, ts, err)
@@ -118,9 +117,9 @@ func TestRetrieveTemplatesByProvider(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/templates", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/templates", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			ts, err := r.RetrieveTemplates()
 			tt.assertFunc(t, ts, err)
@@ -168,9 +167,9 @@ func TestRetrieveTemplateParameters(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/templates/cluster-template/params", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/templates/cluster-template/params", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			ts, err := r.RetrieveTemplateParameters("cluster-template")
 			tt.assertFunc(t, ts, err)
@@ -262,62 +261,11 @@ spec:
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("POST", BaseURI+"/v1/templates/cluster-template/render", tt.responder)
+			httpmock.RegisterResponder("POST", testutils.BaseURI+"/v1/templates/cluster-template/render", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			result, err := r.RenderTemplateWithParameters("cluster-template", nil, capi.Credentials{})
-			tt.assertFunc(t, result, err)
-		})
-	}
-}
-
-func TestCreatePullRequestFromTemplate(t *testing.T) {
-	tests := []struct {
-		name       string
-		responder  httpmock.Responder
-		assertFunc func(t *testing.T, result string, err error)
-	}{
-		{
-			name:      "pull request created",
-			responder: httpmock.NewJsonResponderOrPanic(200, httpmock.File("./testdata/pull_request_created.json")),
-			assertFunc: func(t *testing.T, result string, err error) {
-				assert.Equal(t, result, "https://github.com/org/repo/pull/1")
-			},
-		},
-		{
-			name:      "service error",
-			responder: httpmock.NewJsonResponderOrPanic(500, httpmock.File("./testdata/service_error.json")),
-			assertFunc: func(t *testing.T, result string, err error) {
-				assert.EqualError(t, err, "unable to POST template and create pull request to \"https://weave.works/api/v1/clusters\": something bad happened")
-			},
-		},
-		{
-			name:      "error returned",
-			responder: httpmock.NewErrorResponder(errors.New("oops")),
-			assertFunc: func(t *testing.T, result string, err error) {
-				assert.EqualError(t, err, "unable to POST template and create pull request to \"https://weave.works/api/v1/clusters\": Post \"https://weave.works/api/v1/clusters\": oops")
-			},
-		},
-		{
-			name:      "unexpected status code",
-			responder: httpmock.NewStringResponder(http.StatusBadRequest, ""),
-			assertFunc: func(t *testing.T, result string, err error) {
-				assert.EqualError(t, err, "response status for POST \"https://weave.works/api/v1/clusters\" was 400")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client := resty.New()
-			httpmock.ActivateNonDefault(client.GetClient())
-			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("POST", BaseURI+"/v1/clusters", tt.responder)
-
-			c, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
-			assert.NoError(t, err)
-			result, err := c.CreatePullRequestFromTemplate(capi.CreatePullRequestFromTemplateParams{})
 			tt.assertFunc(t, result, err)
 		})
 	}
@@ -372,9 +320,9 @@ func TestRetrieveCredentials(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/credentials", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/credentials", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			creds, err := r.RetrieveCredentials()
 			tt.assertFunc(t, creds, err)
@@ -408,9 +356,9 @@ func TestRetrieveCredentialsByName(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/credentials", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/credentials", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			creds, err := r.RetrieveCredentialsByName("aws-creds")
 			tt.assertFunc(t, creds, err)
@@ -477,9 +425,9 @@ func TestRetrieveClusters(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/gitops/api/clusters", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/gitops/api/clusters", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			cs, err := r.RetrieveClusters()
 			tt.assertFunc(t, cs, err)
@@ -528,9 +476,9 @@ func TestGetClusterKubeconfig(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/clusters/dev/kubeconfig", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/clusters/dev/kubeconfig", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			k, err := r.GetClusterKubeconfig("dev")
 			tt.assertFunc(t, k, err)
@@ -579,9 +527,9 @@ func TestDeleteClusters(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("DELETE", BaseURI+"/v1/clusters", tt.responder)
+			httpmock.RegisterResponder("DELETE", testutils.BaseURI+"/v1/clusters", tt.responder)
 
-			c, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			c, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			result, err := c.DeleteClusters(clusters.DeleteClustersParams{})
 			tt.assertFunc(t, result, err)
@@ -597,10 +545,10 @@ func TestEntitlementExpiredHeader(t *testing.T) {
 	httpmock.ActivateNonDefault(client.GetClient())
 	defer httpmock.DeactivateAndReset()
 
-	httpmock.RegisterResponder("GET", BaseURI+"/v1/templates", httpmock.ResponderFromResponse(response))
+	httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/templates", httpmock.ResponderFromResponse(response))
 
 	var buf bytes.Buffer
-	c, err := adapters.NewHttpClient(BaseURI, client, &buf)
+	c, err := adapters.NewHttpClient(testutils.BaseURI, client, &buf)
 	assert.NoError(t, err)
 	_, err = c.RetrieveTemplates()
 	assert.NoError(t, err)
@@ -660,9 +608,9 @@ func TestRetrieveTemplateProfiles(t *testing.T) {
 			client := resty.New()
 			httpmock.ActivateNonDefault(client.GetClient())
 			defer httpmock.DeactivateAndReset()
-			httpmock.RegisterResponder("GET", BaseURI+"/v1/templates/cluster-template/profiles", tt.responder)
+			httpmock.RegisterResponder("GET", testutils.BaseURI+"/v1/templates/cluster-template/profiles", tt.responder)
 
-			r, err := adapters.NewHttpClient(BaseURI, client, os.Stdout)
+			r, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
 			assert.NoError(t, err)
 			tps, err := r.RetrieveTemplateProfiles("cluster-template")
 			tt.assertFunc(t, tps, err)
