+++ pkg/capi/capi.go
@@ -23,14 +23,6 @@ type TemplateRenderer interface {
 	RenderTemplateWithParameters(name string, parameters map[string]string, creds Credentials) (string, error)
 }
 
-// TemplatePullRequester defines the interface that adapters
-// need to implement in order to create a pull request from
-// a CAPI template. Implementers should return the web URI of
-// the pull request.
-type TemplatePullRequester interface {
-	CreatePullRequestFromTemplate(params CreatePullRequestFromTemplateParams) (string, error)
-}
-
 // CredentialsRetriever defines the interface that adapters
 // need to implement in order to retrieve CAPI credentials.
 type CredentialsRetriever interface {
@@ -64,20 +56,6 @@ type ProfileValues struct {
 	Name    string `json:"name"`
 	Version string `json:"version"`
 	Values  string `json:"values"`
-}
-
-type CreatePullRequestFromTemplateParams struct {
-	GitProviderToken string
-	TemplateName     string
-	ParameterValues  map[string]string
-	RepositoryURL    string
-	HeadBranch       string
-	BaseBranch       string
-	Title            string
-	Description      string
-	CommitMessage    string
-	Credentials      Credentials
-	ProfileValues    []ProfileValues
 }
 
 type Profile struct {
@@ -208,20 +186,6 @@ func RenderTemplateWithParameters(name string, parameters map[string]string, cre
 	}
 
 	fmt.Fprintf(w, "No template was found.\n")
-
-	return nil
-}
-
-// TODO: extract this to a template pkg
-// CreatePullRequestFromTemplate uses a TemplatePullRequester
-// adapter to create a pull request from a CAPI template.
-func CreatePullRequestFromTemplate(params CreatePullRequestFromTemplateParams, r TemplatePullRequester, w io.Writer) error {
-	res, err := r.CreatePullRequestFromTemplate(params)
-	if err != nil {
-		return fmt.Errorf("unable to create pull request: %w", err)
-	}
-
-	fmt.Fprintf(w, "Created pull request: %s\n", res)
 
 	return nil
 }
