+++ pkg/templates/templates_test.go
@@ -1,0 +1,118 @@
+package templates_test
+
+import (
+	"errors"
+	"net/http"
+	"os"
+	"testing"
+
+	"github.com/go-resty/resty/v2"
+	"github.com/jarcoal/httpmock"
+	"github.com/stretchr/testify/assert"
+	"github.com/weaveworks/weave-gitops/pkg/adapters"
+	"github.com/weaveworks/weave-gitops/pkg/templates"
+	"github.com/weaveworks/weave-gitops/pkg/testutils"
+)
+
+func TestCreatePullRequestFromTemplate_CAPI(t *testing.T) {
+	tests := []struct {
+		name       string
+		responder  httpmock.Responder
+		assertFunc func(t *testing.T, result string, err error)
+	}{
+		{
+			name:      "pull request created",
+			responder: httpmock.NewJsonResponderOrPanic(200, httpmock.File("../adapters/testdata/pull_request_created.json")),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.Equal(t, result, "https://github.com/org/repo/pull/1")
+			},
+		},
+		{
+			name:      "service error",
+			responder: httpmock.NewJsonResponderOrPanic(500, httpmock.File("../adapters/testdata/service_error.json")),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.EqualError(t, err, "unable to POST template and create pull request to \"https://weave.works/api/v1/clusters\": something bad happened")
+			},
+		},
+		{
+			name:      "error returned",
+			responder: httpmock.NewErrorResponder(errors.New("oops")),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.EqualError(t, err, "unable to POST template and create pull request to \"https://weave.works/api/v1/clusters\": Post \"https://weave.works/api/v1/clusters\": oops")
+			},
+		},
+		{
+			name:      "unexpected status code",
+			responder: httpmock.NewStringResponder(http.StatusBadRequest, ""),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.EqualError(t, err, "response status for POST \"https://weave.works/api/v1/clusters\" was 400")
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			client := resty.New()
+			httpmock.ActivateNonDefault(client.GetClient())
+			defer httpmock.DeactivateAndReset()
+			httpmock.RegisterResponder("POST", testutils.BaseURI+"/v1/clusters", tt.responder)
+
+			c, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
+			assert.NoError(t, err)
+
+			result, err := c.CreatePullRequestFromTemplate(templates.CreatePullRequestFromTemplateParams{TemplateKind: templates.CAPITemplateKind})
+			tt.assertFunc(t, result, err)
+		})
+	}
+}
+
+func TestCreatePullRequestFromTemplate_Terraform(t *testing.T) {
+	tests := []struct {
+		name       string
+		responder  httpmock.Responder
+		assertFunc func(t *testing.T, result string, err error)
+	}{
+		{
+			name:      "pull request created",
+			responder: httpmock.NewJsonResponderOrPanic(200, httpmock.File("../adapters/testdata/pull_request_created.json")),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.Equal(t, result, "https://github.com/org/repo/pull/1")
+			},
+		},
+		{
+			name:      "service error",
+			responder: httpmock.NewJsonResponderOrPanic(500, httpmock.File("../adapters/testdata/service_error.json")),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.EqualError(t, err, "unable to POST template and create pull request to \"https://weave.works/api/v1/tfcontrollers\": something bad happened")
+			},
+		},
+		{
+			name:      "error returned",
+			responder: httpmock.NewErrorResponder(errors.New("oops")),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.EqualError(t, err, "unable to POST template and create pull request to \"https://weave.works/api/v1/tfcontrollers\": Post \"https://weave.works/api/v1/tfcontrollers\": oops")
+			},
+		},
+		{
+			name:      "unexpected status code",
+			responder: httpmock.NewStringResponder(http.StatusBadRequest, ""),
+			assertFunc: func(t *testing.T, result string, err error) {
+				assert.EqualError(t, err, "response status for POST \"https://weave.works/api/v1/tfcontrollers\" was 400")
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			client := resty.New()
+			httpmock.ActivateNonDefault(client.GetClient())
+			defer httpmock.DeactivateAndReset()
+			httpmock.RegisterResponder("POST", testutils.BaseURI+"/v1/tfcontrollers", tt.responder)
+
+			c, err := adapters.NewHttpClient(testutils.BaseURI, client, os.Stdout)
+			assert.NoError(t, err)
+			result, err := c.CreatePullRequestFromTemplate(templates.CreatePullRequestFromTemplateParams{TemplateKind: templates.TFTemplateKind})
+			tt.assertFunc(t, result, err)
+		})
+	}
+}
